initial implementation, untested
